reworking the API to support declaration first, then invocation
